+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Breaking changes:
 
 New features:
 
+* (<https://github.com/latk/ganzua/issues/8>) Support `pylock.toml` files as an input format in addition to `uv.lock` and `poetry.lock`.
+
 Fixes:
 
 Other: